move ps report functions
