+++ tslib/src/functions-ps-report/ps-report-3-transformer/mocks/check-config.ts
