+++ tslib/src/functions-ps-report/ps-report-3-transformer/mocks/device.ts
