+++ tslib/src/functions-ps-report/ps-report-4-writer/ps-report-writer.service.ts
@@ -1,6 +1,6 @@
 import { ISqlParameter, ISqlService, SqlService } from '../../sql/sql.service'
 import { ConsoleLogger, ILogger } from '../../common/logger'
-import { IPsychometricReportLine, IReportLineAnswer } from '../../functions/ps-report-3-transformer/models'
+import { IPsychometricReportLine, IReportLineAnswer } from '../ps-report-3-transformer/models'
 import { TYPES, MAX } from 'mssql'
 import * as R from 'ramda'
 
@@ -87,28 +87,28 @@ export class PsReportWriterService {
   private generateSql (data: IPsychometricReportLine): string {
     const updateAnswers = R.repeat(1, 25).map((element, index) => index + 1).map(n => {
       return `
-            Q${n}ID = @q${n}id, Q${n}Response = @q${n}response, Q${n}InputMethods = @q${n}inputMethods, Q${n}K = @q${n}keystrokes, Q${n}Sco = @q${n}score, 
-            Q${n}ResponseTime = @q${n}responseTime, Q${n}TimeOut = @q${n}timeout, Q${n}TimeOutResponse = @q${n}timeoutResponse, 
-            Q${n}TimeOutSco = @q${n}timeoutScore, Q${n}tLoad = @q${n}loadTime, Q${n}tFirstKey = @q${n}firstKey, Q${n}tLastKey = @q${n}lastKey, 
+            Q${n}ID = @q${n}id, Q${n}Response = @q${n}response, Q${n}InputMethods = @q${n}inputMethods, Q${n}K = @q${n}keystrokes, Q${n}Sco = @q${n}score,
+            Q${n}ResponseTime = @q${n}responseTime, Q${n}TimeOut = @q${n}timeout, Q${n}TimeOutResponse = @q${n}timeoutResponse,
+            Q${n}TimeOutSco = @q${n}timeoutScore, Q${n}tLoad = @q${n}loadTime, Q${n}tFirstKey = @q${n}firstKey, Q${n}tLastKey = @q${n}lastKey,
             Q${n}OverallTime = @q${n}overallTime, Q${n}RecallTime = @q${n}recallTime, Q${n}ReaderStart = @q${n}readerStart, Q${n}ReaderEnd = @q${n}readerEnd
       `
     })
     const sql = `
         IF EXISTS ( SELECT * FROM [mtc_results].[psychometricReport] WHERE id = @id )
-        BEGIN        
-           UPDATE [mtc_results].[psychometricReport] 
-           SET 
+        BEGIN
+           UPDATE [mtc_results].[psychometricReport]
+           SET
               PupilId = @pupilId, DOB = @dob, Gender = @gender, Forename = @forename, Surname = @surname, FormMark = @formMark, QDisplayTime = @qDisplayTime,
-              PauseLength = @pauseLength, AccessArr = @accessArr, RestartReason = @restartReason, RestartNumber = @restartNumber, 
+              PauseLength = @pauseLength, AccessArr = @accessArr, RestartReason = @restartReason, RestartNumber = @restartNumber,
               ReasonNotTakingCheck = @ReasonNotTakingCheck, PupilStatus = @pupilStatus, DeviceId = @deviceId, BrowserType = @browserType, SchoolName = @schoolName, Estab = @estab,
               SchoolURN = @schoolURN, LANum = @laNum, AttemptId = @attemptId, FormID = @formId, TestDate = @testDate, TimeStart = @timeStart,
               TimeComplete = @timeComplete, TimeTaken = @timeTaken,
-             
+
               ${updateAnswers.join(',\n')}
-                      
+
            WHERE id = @id;
         END
-        ELSE 
+        ELSE
         BEGIN
           INSERT into [mtc_results].[psychometricReport] (id, DOB, Gender, PupilId, Forename, Surname, FormMark, QDisplayTime, PauseLength,
                                                         AccessArr, RestartReason, RestartNumber, ReasonNotTakingCheck, PupilStatus,
@@ -207,7 +207,7 @@ export class PsReportWriterService {
                 @timeStart,
                 @timeComplete,
                 @timeTaken,
-                                
+
                 @q1Id,
                 @q1response,
                 @q1inputMethods,
@@ -631,8 +631,8 @@ export class PsReportWriterService {
                 @q25overallTime,
                 @q25recallTime,
                 @q25readerStart,
-                @q25readerEnd); 
-        END                                             
+                @q25readerEnd);
+        END
     `
     return sql
   }
