+++ tslib/src/tests-integration/ps-report-writer.service.spec.ts
@@ -1,6 +1,6 @@
 import { PsReportWriterService } from '../functions-throttled/ps-report-4-writer/ps-report-writer.service'
 import { PsychometricReport } from '../functions-throttled/ps-report-4-writer/models'
-import { IPsychometricReportLine } from '../functions/ps-report-3-transformer/models'
+import { IPsychometricReportLine } from '../functions-ps-report/ps-report-3-transformer/models'
 import { SqlService } from '../sql/sql.service'
 import { TYPES } from 'mssql'
 import * as R from 'ramda'
